HW02/requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==2.2.4
pandas>=2.2.3
matplotlib==3.10.0
scikit-learn>=1.15.1
pillow==11.1.0
tqdm>=2.0.0
opencv-python==4.10.0
pycocotools==2.0.8